requirements.txt
matplotlib==3.8.1
numpy==1.26.2
pygame==2.5.2
pygame-widgets==1.1.1
PyYAML==6.0.1
flake8==6.1.0